Done with all services

+++ lib/model/weather_model.dart
@@ -3,13 +3,18 @@ class Weather {
   final double temperature;
   final String mainCondition;
 
-  Weather({required this.name, required this.temperature, required this.mainCondition});
+  Weather(
+      {required this.name,
+      required this.temperature,
+      required this.mainCondition});
 
   factory Weather.fromJson(Map<String, dynamic> json) {
     if (json == null) {
       throw Exception('Failed to parse JSON');
     }
-    if (!json.containsKey('name') || !json.containsKey('main') || !json.containsKey('weather')) {
+    if (!json.containsKey('name') ||
+        !json.containsKey('main') ||
+        !json.containsKey('weather')) {
       throw Exception('Invalid JSON structure');
     }
     if (!json['main'].containsKey('temp') || json['weather'].isEmpty) {
@@ -17,7 +22,7 @@ class Weather {
     }
     return Weather(
       name: json['name'],
-      temperature: json['main']['temp'],
+      temperature: json['main']['temp'].toDouble(),
       mainCondition: json['weather'][0]['description'],
     );
   }
